djangoapp/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.7
django-cors-headers==3.4.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==19.9.0
idna==2.9
psycopg2==2.7.6.1
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
